+++ ghost/core/core/server/data/importer/import-manager.js
@@ -11,7 +11,7 @@ const debug = require('@tryghost/debug')('import-manager');
 const logging = require('@tryghost/logging');
 const errors = require('@tryghost/errors');
 const ImageHandler = require('./handlers/image');
-const ImporterContentFileHandler = require('@tryghost/importer-handler-content-files');
+const ImporterContentFileHandler = require('./handlers/ImporterContentFileHandler');
 const RevueHandler = require('./handlers/revue');
 const JSONHandler = require('./handlers/json');
 const MarkdownHandler = require('./handlers/markdown');
